mseek must used long int for compatibilities with 64 bits OS

+++ scilab/modules/fileio/src/cpp/mseek.cpp
@@ -26,7 +26,7 @@ extern "C"
 }
 /*--------------------------------------------------------------------------*/
 
-int mseek(int fd, int offset, int iflag)
+int mseek(int fd, long int offset, int iflag)
 {
 #if (defined(sun) && !defined(SYSV)) || defined(sgi)
     int irep;
@@ -63,12 +63,12 @@ int mseek(int fd, int offset, int iflag)
 #else
 #ifdef _MSC_VER
 #if _WIN64
-    if (_fseeki64(fa, (long) offset, iflag) == -1 )
+    if (_fseeki64(fa, offset, iflag) == -1 )
 #else
-    if (fseek(fa, (long) offset, iflag) == -1 )
+    if (fseek(fa, offset, iflag) == -1 )
 #endif
 #else
-    if (fseek(fa, (long) offset, iflag) == -1)
+    if (fseek(fa, offset, iflag) == -1)
 #endif
     {
         return errno; // global variable produced by fseek
